I have refactored the entire module completely for task #1287. I will discuss a pull request tomorrow
