+++ src/playerclass/CMakeLists.txt
@@ -2,6 +2,7 @@ add_library(playerclass
         src/class.c
         src/class_item.c
         src/class_prefabs.c
+        src/multiclass.c
         )
 
 target_include_directories(playerclass PRIVATE src/)
